+++ tests/Persisters/FilesystemTest.php
@@ -22,7 +22,7 @@ class FilesystemTest extends TestCase
 
         $this->persistable = new DummyClassifier();
 
-        $this->persister = new Filesystem($this->path, 0);
+        $this->persister = new Filesystem($this->path, 1);
     }
 
     public function test_build_persister()
@@ -31,18 +31,25 @@ class FilesystemTest extends TestCase
         $this->assertInstanceOf(Persister::class, $this->persister);
     }
 
-    public function test_save_load_and_delete()
+    public function test_save_and_load()
     {
         $this->assertFalse(file_exists($this->path));
 
+        $this->persister->save($this->persistable);
         $this->persister->save($this->persistable);
 
         $this->assertFileExists($this->path);
 
         $model = $this->persister->load();
+        $old = $this->persister->load(1);
 
         $this->assertInstanceOf(DummyClassifier::class, $model);
         $this->assertInstanceOf(Persistable::class, $model);
+
+        $this->assertInstanceOf(DummyClassifier::class, $old);
+        $this->assertInstanceOf(Persistable::class, $old);
+
+        $this->persister->flush();
 
         unlink($this->path);